recursive selection from mvh/i18nMenusAgain + add en back if manually removed
