+++ package.json
@@ -10,7 +10,7 @@
     "dompurify": "^3.2.4",
     "eslint-config-react-app": "^7.0.1",
     "font-detect-rhl": "^1.1.7",
-    "pithekos-lib": "^0.5.11",
+    "pithekos-lib": "^0.5.24",
     "prop-types": "^15.8.1",
     "react": "^17.0.0 || ^18.0.0",
     "react-dom": "^17.0.0 || ^18.0.0"
